requirements.txt
streamlit
pandas
openai
python-dotenv
faiss-cpu
